src/main/java/com/task/lecturesschedule/service/impl/StudentServiceImpl.java
package com.task.lecturesschedule.service.impl;

import com.task.lecturesschedule.model.Student;
import com.task.lecturesschedule.repository.StudentRepository;
import com.task.lecturesschedule.service.StudentService;
import java.util.List;
import org.springframework.stereotype.Service;

@Service
public class StudentServiceImpl implements StudentService {
    private final StudentRepository studentRepository;

    public StudentServiceImpl(StudentRepository studentRepository) {
        this.studentRepository = studentRepository;
    }

    @Override
    public Student add(Student student) {
        return studentRepository.save(student);
    }

    @Override
    public Student get(Long id) {
        return studentRepository.getById(id);
    }

    @Override
    public List<Student> getAll() {
        return studentRepository.findAll();
    }

    @Override
    public Student update(Student student) {
        return studentRepository.save(student);
    }

    @Override
    public void delete(Long id) {
        studentRepository.delete(get(id));
    }
}
